Move getUserName helper to shared
- in preparation for Workplace Search plugin also using this helper
