+++ x-pack/plugins/enterprise_search/public/applications/shared/get_username/get_username.ts
